requirements.txt
fastapi
uvicorn
biopython
psutil
python-multipart
email_validator
